// src/lib/resources/fates.data.ts
type FateData = {
  readonly id: number
  readonly name: string
  readonly icon: string
  readonly eureka: number
  readonly rule: number
  readonly specialFate: boolean
  readonly x: number
  readonly y: number
  readonly z: number
  readonly zoneId: number
}